+++ frontend/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.0.1",
-    "@types/axios": "^0.14.0",
     "@types/react-redux": "^7.1.33",
     "@types/react-slick": "^0.23.13",
     "axios": "^1.6.2",
